requirements.txt
discord.py
matplotlib

